+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2026-04-30
+
 ### Added
 
 - `WildFlyImageRegistry::load_or_update()` and `FeaturePackRegistry::load_or_update()` convenience methods that automatically download the configuration if it is missing or corrupt, retrying once on load failure
@@ -54,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release 🎉
 
-[Unreleased]: https://github.com/hpehl/wildfly-meta/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/hpehl/wildfly-meta/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/hpehl/wildfly-meta/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/hpehl/wildfly-meta/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/hpehl/wildfly-meta/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/hpehl/wildfly-meta/compare/v0.1.0...v0.2.0